edge-client-sdk/pyproject.toml
[project]
name = "edge-client-sdk"
version = "0.1.0"
description = "Edge-side supervision modules extracted from yolodemo and aligned to shared contracts."
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "numpy>=2.2.4",
  "opencv-python-headless>=4.12.0.88",
  "pydantic>=2.11.1",
  "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
  "pytest>=8.3.5",
]

[tool.uv]
package = true
default-groups = ["dev"]

[project.scripts]
edge-replay-demo = "edge.uplink.replay_client:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
